refactor(architecture): sink the shared webview icons into the kernel
src/webview/webviewIcons.ts -> src/webviewIcons.ts (MOD-SHARED-KERNEL): the
pure SVG constants sat in the composition top, so every feature webview
imported upwards. The move breaks the DASHBOARD-SHELL<->PROMPT-LIBRARY and
AI-SESSION-CONVERSATION<->DASHBOARD-SHELL 2-cycles and ejects the prompt
library from the module SCC cluster (12 -> 11 modules). Consumers, check
scripts, and the behavior-contract evidence path move to the new location;
the icon bytes are unchanged and the prompt library entrypoints narrow to
the consumed surface.
