+++ .ci/architecture-debt-baseline.json
@@ -8,15 +8,13 @@
         "2:MOD-AI-SESSION-ATTENTION->MOD-AI-SESSION-PROVIDER",
         "2:MOD-AI-SESSION-ATTENTION->MOD-AI-SESSION-RUNTIME",
         "2:MOD-AI-SESSION-ATTENTION->MOD-WORKSPACE-IDENTITY",
-        "2:MOD-AI-SESSION-CONVERSATION->MOD-DASHBOARD-SHELL",
         "2:MOD-AI-SESSION-PROVIDER->MOD-AI-SESSION-RUNTIME",
         "2:MOD-AI-SESSION-PROVIDER->MOD-DASHBOARD-SHELL",
         "2:MOD-AI-SESSION-PROVIDER->MOD-WORKSPACE-IDENTITY",
         "2:MOD-DASHBOARD-SHELL->MOD-OPEN-WORKSPACE",
-        "2:MOD-DASHBOARD-SHELL->MOD-PROMPT-LIBRARY",
         "2:MOD-DASHBOARD-SHELL->MOD-SKILL-MANAGEMENT",
         "2:MOD-DASHBOARD-SHELL->MOD-TODO",
-        "scc:MOD-AI-SESSION-ATTENTION|MOD-AI-SESSION-CONTROL|MOD-AI-SESSION-CONVERSATION|MOD-AI-SESSION-PROVIDER|MOD-AI-SESSION-RUNTIME|MOD-DASHBOARD-SHELL|MOD-OPEN-WORKSPACE|MOD-PROJECT-CATALOG|MOD-PROMPT-LIBRARY|MOD-SKILL-MANAGEMENT|MOD-TODO|MOD-WORKSPACE-IDENTITY"
+        "scc:MOD-AI-SESSION-ATTENTION|MOD-AI-SESSION-CONTROL|MOD-AI-SESSION-CONVERSATION|MOD-AI-SESSION-PROVIDER|MOD-AI-SESSION-RUNTIME|MOD-DASHBOARD-SHELL|MOD-OPEN-WORKSPACE|MOD-PROJECT-CATALOG|MOD-SKILL-MANAGEMENT|MOD-TODO|MOD-WORKSPACE-IDENTITY"
       ]
     }
   }
