+++ docs/testing/architecture-modules.json
@@ -101,7 +101,6 @@
       "mayDependOn": [
         "MOD-AI-SESSION-PROVIDER",
         "MOD-AI-SESSION-RUNTIME",
-        "MOD-DASHBOARD-SHELL",
         "MOD-SHARED-KERNEL",
         "MOD-WORKTREE-LIFECYCLE"
       ],
@@ -787,10 +786,13 @@
         "exclude": []
       },
       "publicEntrypoints": [
-        "src/prompts/**"
+        "src/prompts/dashboardController.ts",
+        "src/prompts/service.ts",
+        "src/prompts/terminalCommandController.ts",
+        "src/prompts/webviewContent.ts"
       ],
       "mayDependOn": [
-        "MOD-DASHBOARD-SHELL"
+        "MOD-SHARED-KERNEL"
       ],
       "roles": [
         {
@@ -881,7 +883,8 @@
           "src/constants.ts",
           "src/sessionAssignment.ts",
           "src/worktreeSessionAssignment.ts",
-          "src/worktreeIdentity.ts"
+          "src/worktreeIdentity.ts",
+          "src/webviewIcons.ts"
         ],
         "exclude": []
       },
@@ -890,7 +893,8 @@
         "src/constants.ts",
         "src/sessionAssignment.ts",
         "src/worktreeSessionAssignment.ts",
-        "src/worktreeIdentity.ts"
+        "src/worktreeIdentity.ts",
+        "src/webviewIcons.ts"
       ],
       "mayDependOn": [
         "MOD-AI-SESSION-PROVIDER",
@@ -907,7 +911,8 @@
             "src/constants.ts",
             "src/sessionAssignment.ts",
             "src/worktreeSessionAssignment.ts",
-            "src/worktreeIdentity.ts"
+            "src/worktreeIdentity.ts",
+            "src/webviewIcons.ts"
           ]
         }
       ],
